+++ app/models/carto/user_multifactor_auth.rb
@@ -17,9 +17,14 @@ module Carto
     validates_uniqueness_of :type, scope: :user_id
 
     before_create :create_shared_secret
+    after_save    :sync_central
+    after_destroy :sync_central
 
     self.inheritance_column = :_type
 
+    scope :enabled, -> { where(enabled: true) }
+    scope :setup, -> { where(enabled: false) }    
+    
     def self.new_from_hash(uma_hash)
       new(
         created_at: uma_hash[:created_at],
@@ -56,6 +61,10 @@ module Carto
     end
 
     private
+
+    def sync_central
+      ::User[user.id].update_in_central
+    end
 
     def last_login_in_seconds
       last_login.strftime('%s').to_i if last_login